requirements.txt
pyaudio==0.2.13
numpy==1.23.4
flask==2.3.2
requests==2.30.0
openai-whisper==20230314
TTS==0.13.3
transformers[torch]==4.29.2
